Se comenta el formulario de django-bootstrap3 porque actualmente esta generando conflictos
