+++ biodiversidad_app/templates/biodiversidad_app/_elements/_modal-sign-up.html
@@ -6,7 +6,8 @@
                 <h3 class="text-center modal-title">Registro</h3>
             </div>
             <div class="modal-body">
-                {% include 'biodiversidad_app/_forms/user_registration.html' %}
+                {#% include 'biodiversidad_app/_forms/user_registration.html' %#}
+                {% include 'biodiversidad_app/_forms/registro_usuario.html' %}
             </div>
         </div>
     </div>
