+++ biodiversidad_miso/local_settings.py
@@ -3,11 +3,10 @@ import os
 import dj_database_url
 
 # Please use this dict when deploying to heroku
-'''
 DATABASE_DICT = dj_database_url.config(default=os.environ.get('DATABASE_URL', ''))
-'''
-# Please use this dict for locale tests
 
+# Please use this dict for locale tests
+'''
 DATABASE_DICT = {
     'ENGINE': 'django.db.backends.postgresql',
     'NAME': 'biodiversidad_db',
@@ -16,4 +15,4 @@ DATABASE_DICT = {
     'HOST': '127.0.0.1',
     'PORT': '5432',
 }
-
+'''
