+++ biodiversidad_miso/settings.py
@@ -38,8 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'bootstrap3',
     'biodiversidad_app.apps.BiodiversidadAppConfig',
-    'bootstrap3'
 ]
 
 MIDDLEWARE = [
